+++ storybook/stories/settings/general/General.stories.js
@@ -16,6 +16,7 @@ import { locales, themesIds } from '../../_support/config';
 
 // Screens
 import ProfileSettingsForm from '../../../../source/renderer/app/components/widgets/forms/ProfileSettingsForm';
+import StakePoolsSettings from '../../../../source/renderer/app/components/settings/categories/StakePoolsSettings';
 import DisplaySettings from '../../../../source/renderer/app/components/settings/categories/DisplaySettings';
 import SupportSettings from '../../../../source/renderer/app/components/settings/categories/SupportSettings';
 import TermsOfUseSettings from '../../../../source/renderer/app/components/settings/categories/TermsOfUseSettings';
@@ -45,6 +46,15 @@ storiesOf('Settings|General', module)
       currentLocale={LANGUAGE_OPTIONS[0].value}
       currentNumberFormat={NUMBER_OPTIONS[0].value}
       currentTimeFormat={TIME_OPTIONS[0].value}
+    />
+  ))
+  .add('Stake Pools', () => (
+    <StakePoolsSettings
+      onSelectSmashServerUrl={action('onSelectSmashServerUrl')}
+      onResetSmashServerError={action('onResetSmashServerError')}
+      isLoading={false}
+      smashServerUrl="https://smash.cardano-mainnet.iohk.io"
+      onOpenExternalLink={action('onOpenExternalLink')}
     />
   ))
   .add('Themes', () => (